move analFunctionSignature to anal
